# Fix casting nan and infinity to floating-point types