--- a/velox/expression/CMakeLists.txt
+++ b/velox/expression/CMakeLists.txt
@@ -50,8 +50,14 @@ add_library(
   VectorFunction.cpp)
 
 target_link_libraries(
-  velox_expression velox_core velox_vector velox_common_base
-  velox_expression_functions velox_functions_util)
+  velox_expression
+  velox_core
+  velox_vector
+  velox_common_base
+  velox_expression_functions
+  velox_functions_util
+  double-conversion::double-conversion
+  Folly::folly)
 
 add_subdirectory(type_calculation)
 add_subdirectory(signature_parser)