--- a/velox/expression/CastHooks.h
+++ b/velox/expression/CastHooks.h
@@ -22,8 +22,8 @@
 namespace facebook::velox::exec {
 
 /// This class provides cast hooks to allow different behaviors of CastExpr and
-/// SparkCastExpr. The main purpose is crate customized cast implementation by
-/// taking full usage of existing cast expression.
+/// SparkCastExpr. The main purpose is to create customized cast implementation
+/// by taking full usage of existing cast expression.
 class CastHooks {
  public:
   virtual ~CastHooks() = default;
@@ -33,6 +33,14 @@ class CastHooks {
 
   virtual Expected<int32_t> castStringToDate(
       const StringView& dateString) const = 0;
+
+  // 'data' is guaranteed to be non-empty and has been processed by
+  // removeWhiteSpaces.
+  virtual Expected<float> castStringToReal(const StringView& data) const = 0;
+
+  // 'data' is guaranteed to be non-empty and has been processed by
+  // removeWhiteSpaces.
+  virtual Expected<double> castStringToDouble(const StringView& data) const = 0;
 
   // Returns whether legacy cast semantics are enabled.
   virtual bool legacy() const = 0;