--- a/velox/expression/PrestoCastHooks.cpp
+++ b/velox/expression/PrestoCastHooks.cpp
@@ -14,8 +14,14 @@
  * limitations under the License.
  */
 
+#include <cmath>
+
+#include <double-conversion/double-conversion.h>
+#include <folly/Expected.h>
+
 #include "velox/expression/PrestoCastHooks.h"
 #include "velox/external/date/tz.h"
+#include "velox/functions/lib/string/StringImpl.h"
 #include "velox/type/TimestampConversion.h"
 
 namespace facebook::velox::exec {
@@ -66,6 +72,57 @@ Expected<int32_t> PrestoCastHooks::castStringToDate(
   // Cast from string to date allows only complete ISO 8601 formatted strings:
   // [+-](YYYY-MM-DD).
   return util::fromDateString(dateString, util::ParseMode::kPrestoCast);
+}
+
+namespace {
+
+using double_conversion::StringToDoubleConverter;
+
+template <typename T>
+Expected<T> doCastToFloatingPoint(const StringView& data) {
+  static const T kNan = std::numeric_limits<T>::quiet_NaN();
+  static StringToDoubleConverter stringToDoubleConverter{
+      StringToDoubleConverter::ALLOW_TRAILING_SPACES,
+      /*empty_string_value*/ kNan,
+      /*junk_string_value*/ kNan,
+      "Infinity",
+      "NaN"};
+  int processedCharactersCount;
+  T result;
+  auto* begin = std::find_if_not(data.begin(), data.end(), [](char c) {
+    return functions::stringImpl::isAsciiWhiteSpace(c);
+  });
+  auto length = data.end() - begin;
+  if (length == 0) {
+    // 'data' only contains white spaces.
+    return folly::makeUnexpected(Status::UserError());
+  }
+  if constexpr (std::is_same_v<T, float>) {
+    result = stringToDoubleConverter.StringToFloat(
+        begin, length, &processedCharactersCount);
+  } else if constexpr (std::is_same_v<T, double>) {
+    result = stringToDoubleConverter.StringToDouble(
+        begin, length, &processedCharactersCount);
+  }
+  // Since we already removed leading space, if processedCharactersCount == 0,
+  // it means the remaining string is either empty or a junk string. So return a
+  // user error in this case.
+  if UNLIKELY (processedCharactersCount == 0) {
+    return folly::makeUnexpected(Status::UserError());
+  }
+  return result;
+}
+
+} // namespace
+
+Expected<float> PrestoCastHooks::castStringToReal(
+    const StringView& data) const {
+  return doCastToFloatingPoint<float>(data);
+}
+
+Expected<double> PrestoCastHooks::castStringToDouble(
+    const StringView& data) const {
+  return doCastToFloatingPoint<double>(data);
 }
 
 StringView PrestoCastHooks::removeWhiteSpaces(const StringView& view) const {