Summary: 
Presto only allow casting "Infinity" and "NaN" to floating-point types, 
but not these strings in different letter cases or "Inf". Velox allows 
these strings to be case insensitive. Since Spark cast function allows 
different letter cases of "NaN", "Infinity", and "Inf", this diff adds APIs 
of casting string to REAL and DOUBLE to CastHooks and fixes Velox's 
Presto cast function in PrestoCastHooks.

Before the fix:
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast_varchar_as_double##cast_valid                        108.03ms      9.26
cast_varchar_as_double##try_cast_invalid_nan               51.91ms     19.26
cast_varchar_as_double##try_cast_invalid_infini            53.39ms     18.73
```

After the fix:
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast_varchar_as_double##cast_valid                         85.87ms     11.65
cast_varchar_as_double##try_cast_invalid_nan               35.05ms     28.53
cast_varchar_as_double##try_cast_invalid_infini            71.88ms     13.91
```

Differential Revision: D57888059